requirements.txt
langchain
langchain-core
langchain-community
pymupdf
sentence-transformers
chromadb
langchain-groq
fastapi
uvicorn
python-dotenv